requirements.txt
fastapi
uvicorn[standard]
websockets
opencv-python
numpy
onvif-zeep
pymysql
sqlalchemy
pytz
python-dotenv